blueprints: Convert blueprint tests to integration tests
mocking out a whole server's worth of responses really only tests the
mock code, not the actual behavior of the application. Only integration
testing against an actual server can do that.

So this changes the blueprints_test code to use a real server, which is
setup each time the tests are run for consistent results. Run 'make
integration' to create composer-cli-tests and then run that binary on a
system with osbuild-composer installed.

Currently the TestUndoMissingCommit fails, this is because
osbuild-composer returns the wrong error type when the undo commit isn't
found.