weldr: Add new GetJSONAllFnTotal function
This function takes a function that returns the total count to use when
requesting json data. Most of the paginated responses have total at the
top level, which GetJSONAll continues to handle. But some of them, like
blueprints/changes, have more than 1 total value, or it isn't at the top
level. This allows a function to be passed that will parse the response
body and return the total that will return the most results.