composer-cli: Update the json output with more details
The output generated by `--json` now includes the method, the path, the
status code, and the json body of the response from the server. For
example The output of composer-cli --json blueprints list is:

{
    "method": "GET",
    "path": "/blueprints/list?limit=0",
    "status": 200,
    "body": {
        "blueprints": [],
        "limit": 0,
        "offset": 0,
        "total": 4
    }
}
{
    "method": "GET",
    "path": "/blueprints/list?limit=4",
    "status": 200,
    "body": {
        "blueprints": [
            "test-blueprint-1",
            "test-blueprint-2",
            "test-blueprint-3",
            "test-blueprint-4",
        ],
        "limit": 4,
        "offset": 0,
        "total": 4
    }
}

Some requests send multiple requests in order to retrieve all of the
responses, this is one example of that behavior.