tests: Fix integration tests
osbuild-composer v31 returns a more detailed error for
TestStartComposeBadType, so that has been changed.

The number of image types returned depends on the distribution being
used, so change to looking for >= 6 of them.

The source type depends on the first repo in the list, so it could be
any of the supported types. They all start with "yum-" so change the
test to look for that instead.