composer-cli: Return a helpful error for socket problems
Tell the user to start osbuild-composer if the socket file is missing,
and tell them to check the socket's group (usually weldr, but it pulls
it from the socket file itself) if the file exists but cannot be
accessed.

Includes test for present and missing socket file.
We cannot test for permissions because as root it always has access, and
as a user it cannot make a file that it doesn't already have access to.