cmdline: Add package glob support to modules list command
The API supports searching for packages using '*' globs. This adds
support for using that via 'composer-cli modules list foo bar* baz' to
search for packages matching the exact name or wildcard.

This includes both unit tests for the cmdline and integration tests for
the weldr client library which has a new SearchModules function to
support this.