saveBlueprint: Return the filename to the caller
The function constructs the filename from the blueprint name, optional
commit, and path. Return it to the caller so that it can be used to
access the file.

This also drops the previous checks for /, .., and . which can never be
returned because the blueprint name is processed by filepath.Base, but
it does check for '.toml' which would indicate a blank blueprint name,
and returns an error.