Add support for compose start warning messages
osbuild-composer v79 can include a list of warning strings in the
compose start result structure. This commit adds support for that to
weldr/compose.go and to composer-cli which will print 'Warning:
<string>' for each of them before printing the new compose UUID.

ComposeStartV0 has been changed to add a Warnings list of strings. Since
this is an addition I didn't bump the API version. Callers should be
testing the response Status flag to determine if the request was
successful. It may now also contain Warnings when the flag is true.

The startCompose functions can return an APIResponse that includes the
Warnings while also returning the UUID. In the future it may be possible
to return errors and warnings, so the APIResponse Status flag should
continue to be used to determine if the response is an error (Status ==
false).