compose: Add compose wait command
This command is used to wait for a compose to be finished, fail, or time
out.

The `--timeout` option can be used to change the default timeout which
is 5m. The string is parsed using https://pkg.go.dev/time#ParseDuration

`--poll` can be used to set the polling interval.