compose: Add support for cloud API UUIDs to compose wait
The composes built with the cloud API are not stored in the weldr API
store so the UUID from the user may belong to one or the other. Check
cloud API first, then weldr API and return the weldr API error if it is
not found in either place.

Note that the result printed is different. Cloud API uses 'success' and
'failure' and WELDR API uses 'FINISHED' and 'FAILED'.

Related: RHEL-60124
Related: RHEL-60134