lint: Clean up lint errcheck errors
The linter complains about unchecked errors when deferring Close which
in most cases makes no sense to check (nothing can be done if there is
an error). This flags them with nolint:errcheck

In some other places, mostly in testing, an error on close could be
useful when debugging a failing test, so I added assert.Nil to those.